resources/python-backend/pyproject.toml
[project]
name = "elato-server"
version = "0.1.0"
description = "ELATO - Open Toys for All"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "ElatoAI, Akashdeep Deb" }]

dependencies = [
    "mlx==0.31.0",
    "mlx-lm==0.31.0",
    "mlx-vlm==0.4.0",
    "mlx-audio[stt,tts]==0.4.0",
    "einops>=0.8.1",
    "transformers==5.1.0",
    "librosa",
    "pyloudnorm",
    "soundfile",
    "tiktoken",
    "soxr",
    "numpy",
    "numba",
    "webrtcvad-wheels==2.0.14",
    "uvicorn[standard]==0.40.0",
    "fastapi==0.127.0",
    "huggingface-hub==1.6.0",
    "esptool==5.2.0",
    "pyserial==3.5",
    "av>=16.0.0",
    "zeroconf>=0.136.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
